demo/auth-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
httpx==0.25.1
